+++ examples/mask.rs
@@ -12,8 +12,8 @@ fn main() {
         .transform(Transform::from_row(1.0, -0.3, 0.0, 1.0, 0.0, 75.0))
         .unwrap();
 
-    let mut mask = Mask::new();
-    mask.set_path(500, 500, &clip_path, FillRule::EvenOdd, true);
+    let mut mask = Mask::new(500, 500).unwrap();
+    mask.fill_path(&clip_path, FillRule::EvenOdd, true, Transform::default());
 
     let mut paint = Paint::default();
     paint.set_color_rgba8(50, 127, 150, 200);