+++ src/lib.rs
@@ -58,7 +58,7 @@ mod painter; // Keep it under `pixmap` for a better order in the docs.
 pub use blend_mode::BlendMode;
 pub use color::{Color, ColorU8, PremultipliedColor, PremultipliedColorU8};
 pub use color::{ALPHA_OPAQUE, ALPHA_TRANSPARENT, ALPHA_U8_OPAQUE, ALPHA_U8_TRANSPARENT};
-pub use mask::Mask;
+pub use mask::{Mask, MaskType};
 pub use painter::{FillRule, Paint};
 pub use pixmap::{Pixmap, PixmapMut, PixmapRef, BYTES_PER_PIXEL};
 pub use shaders::{FilterQuality, GradientStop, PixmapPaint, SpreadMode};